fix teardown bug
I'm not completely sure of the mechanism, but by assigning the value of
`builtins.__xonsh_env__.get('VI_MODE')` to an instance attribute the PT
history teardown routine throws an error on exit.

Instead, just feed the True/False value from the `.get()` directly to
the lambda function that runs the vi_mode keybinding.  Everything still
works as expected and no errors on exit.