fix some completer.py lint errors
This was originally https://github.com/scopatz/xonsh/pull/508 but I screwed
up my branch. Since I also needed to do more restructuring of the code in
light of feedback from @scopatz I decided to save the diff, close that
PR, recreate the branch, then reapply the diff and continue working on it
to reach this state.