add space escaping to bash completion function
On Linux, if there's a foldername with a space in it, like "bad folder"
then tab-completion using `cd` will wrap single quotes around the name,
e.g.

```
cd bad(TAB)
```

completes to

```
cd 'bad folder/'
```

This behavior is encoded in `path_complete` in `completer.py`.  But if
you want to remove that poorly named folder, the argument to `rm` is
completed by `bash_complete` which didn't have this tweak and so the
result is

```
rm -r bad(TAB)
```
leads to

```
rm -r bad folder
/usr/bin/rm: cannot remove 'bad': No such file or directory
/usr/bin/rm: cannot remove 'folder/': No such file or directory
```

This tweaks the `bash_complete` function so that

```
rm -r bad(TAB)
```

returns

```
rm -r 'bad folder/'
```