add set of completion wrap tokens for bash escape
I've noticed a few more problems in the way that the xonsh parser can
interfere with bash completion (or vice versa).  This adds a set of
tokens that might show up in filenames that cause trouble.  If one of
these tokens is in a filename/path then the bash completer will return
the `repr` of the filename and the xonsh parser won't try to interpret
the special symbols in the filename