Eliminate `locate_binary` and use a native Python implementation instead
We already had a Python implementation for `which`/`where`
functionality but for some reason we were not making enough use of it.
This patch moves ` built_ins._get_runnable_name` to
`environ.locate_binary` and makes everything use it. This improves `git`
and `hg` querying speeds for prompt on Windows.