fix collision with ptk multiline dedent and `pass`
This is a bit of an edge case, but the password manager `pass` causes a
few headaches with the multiline carriage return logic.  Specifically,
because `pass` is a `DEDENT_TOKEN` when running `pass` as the password
manager the prompt will delete one indent-length's worth of characters
on a `<ENTER>` press.  This is a one-line fix to make sure that people
 can use `pass` in the PTK shell.
Just running `pass` will be interpreted as a Pythonic `pass`, but
commands like `pass git init` will work.

If someone has a xsh script that uses `pass` it should be wrapped in
`$[]`.