fix ptk completer to allow completion earlier in line
As reported by @asmeurer, the completer had incorrect index values for
completion when the cursor was located before a space, that is

`~/Do<C>emacsclient -nw`

would not offer file path completions (where the cursor is marked by
`<C>`). This was because of (my) naive searching for indices, now
slightly less naive.