add mock xonsh env to test
`test_format_prompt_with_invalid_func` raises an exception
(intentionally).  when this is checked against `partial_format_prompt`
it eventually makes its way to `print_exception` in `tools.py` which
expects `env` to have an attribute `is_manually_set`.

This just adds a mock xonsh env to the test so that `env` is properly
defined.