enable character selection in vi_mode
When `open_in_editor` is enabled, the `VI_MODE` binding to open the
editor is `v`, which is unexpected, as `v` should be character
selection.

The `open_in_editor` option has been disabled, but the `Ctrl-X Ctrl-E`
shortcut to open the current buffer in an editor has been explicitly
enabled, so there shouldn't be any change of behavior for anyone except
those who were using `v` to open an editor in `VI_MODE` (which I think
is no one).