stop autocomplete from autoselecting first entry
This was mentioned in #1338.  If I have the following directory
structure

```console
foo/
 -bar/
 -bag/
 -bing/
 -bong/
```

and have already completed to `cd foo/` and hit TAB, ptk will autoselect
the first entry of the completions.  This is incredibly annoying if
there are many possible selections and you would like to filter through
them by prefix.

This forks the `generate_completions` event from PTK and stops it from
autoselecting the first entry (but still displays possibilities)