add a 'kill' alias to use jobs indexes
The zsh jobs management allow user to call a 'kill' builtin command with
'%' + index of the job, to kill a given job. This patch introduces the
same behavior in xonsh.

Unfortunately, as it calls the system 'kill' command, which doesn't exist
under Windows, I've added this alias only for unix systems.

A better way would be to re-implement the kill binary, and calling the
os.kill() function.