split subproc parser into a specific parser
The goal of this patch is to allow an external program depending on xonsh
as a library to use Shell with the Python parser, but without subprocess
commands.

So it splits the current BaseParser into two classes, and a "PythonParser"
class to be used as base for another parser class.

It is also possible to give a class object to Shell and Execer
cunstructors.