Use ansi color names for default style
This updates XONSH_BASE_STYLE to match the default style in pygments.
Previously, XONSH_BASE_STYLE was only use if pygments was not installed.
Now, XONSH_BASE_STYLE is always used if style is set to 'default'.
In case pygment 2.2 is available a version style with color names
is used (#ansidarkblue etc.).