checkout HEAD before attempting second deploy
I think the errors that doctr encounters when pushing the second set of
docs has to do with trying to add to the gh-pages branch when we're
starting from non-HEAD location?

I'm hoping that by manually putting us back onto HEAD that this should
avoid that...

If this does work, it's probably worth putting into doctr itself.

Usually caveats that we can't test this until it's merged...