use ptk-specific `push` when executing commands
When events were introduced, the PTK shell's `_push` was left behind.
This is the cause of #2036 as the buffer isn't cleared between separate
prompt inputs, leading to very, very strange behavior.

This restores the use of the PTK `_push` that properly resets the buffer
if the AST fails to compile and also adds in the precommand src
transformation so that the bashisms xontrib still works as expected.