Build a background predictor from analyzing the binary file.
This commit introduce a predictor generator from analyzing the binary of the
executable.
It works for emacs, weechat, vim, nano, mutt for example.
The main idea is to detect the use of ncurses or manual tty handling with
isatty, tcgetattr, and tcgetattr.
It should works on POSIX, fully tested on Linux, predictor generator function is
also tested on MacOS.