remove tokenize from pyghooks
Moving anything that doesn't explicitly require Pygments to its own
file. We still need Tokens, but we don't need Pygments tokens.  Instead,
I've added in the simple Token class from PTK, but it's vended so we
don't need to perform a check as to whether pygments or prompt_toolkit
is available before running