Fix the regression introduced by #2235 concerning the `$PWD` change
Before 72d04e (#2235), the `_fix_cwd` function was having a side effect to reset
the `$PWD` environment variable if it was deleted. As #2235 changed this
behviour in order to handle symlinks, this implicit side effect is gone.

This commit reintroduce the reset of the `$PWD` environment variable from
`_fix_cwd`, but explicitely so it wont be missed again :-)

    «Explicit is better than implicit.» [Tim Peters]

Close #2236