Introduce deprecated function/method decorator
This method allows for import time modification of a decorated
method/function such that it's docstring will reflect when the
function was deprecated, when it will be removed and that it is in
fact deprecated. It will also use the warnings system to flag
a DeprecationWarning that Python will report on when run with the
`-W` flag, or, when run through certain static/dynamic code
analyzers.