raise AssertionError on deprecated when removed_in exceeded
As there was some worry raised over whether people would remember
to remove a deprecated function once the removed_in version had been
reached, I've decided to add functionality that will force the
issue. When a @deprecated function is tagged with a removed_in
release that is greater than or equal to the current release, the
@deprecated decorator will raise an AssertionError, which will
hopefully break CI/CD and force the function to be removed.