gistatus: Fix hash not being shown when in detaced HEAD and there are no tags.
The `git describe --exact-match` used until fails with
  fatal: No names found, cannot describe anything.
if there are no tags at all in the repository.
In that case, it exits with exit code 128, and thus the
control flow was aborted, and the fallback `git rev-parse`
was not executed.

As a workaround, `git describe --always` can be used,
which achieves the intended result without increasing the
number of `git` invocations.
See also:
  https://stackoverflow.com/questions/4916492/git-describe-fails-with-fatal-no-names-found-cannot-describe-anything