Add history matching toggle to search entire line
Default behavior of prompt_toolkit on an up-arrow history search is to use a
`startswith` check to match previous entries.
This introduces an environment variable `XONSH_HISTORY_MATCH_ANYWHERE` that, if
`True`, will instead perform a history search with an `is in` to allow matching
anywhere.

This could potentially have performance impacts depending on how many history
items are being searched, so I've defaulted it to `False` but I'm open to input
on that (or any of the rest of it).