Implement VIRTUAL_ENV_PROMPT and _DISABLE_PROMPT
Fully implements the suppression of virtualenv name display
when `$VIRTUAL_ENV_DISABLE_PROMPT` is defined and truthy.

Implements the xonsh-side "catch" of a custom prompt defined
via the `--prompt` argument to `virtualenv`. In order for this to
work, `activate.xsh` will need to be modified on the virtualenv
side to define `$VIRTUAL_ENV_PROMPT` during activate, if the
environment is created with `--prompt="foo"`.

Partially implements pypa/virtualenv#1299