Implement more bang commands: !$ !* !^ !<cmd>
While the 'bashisms' xontrib provides the !! command, related commands
like !$, !*, !^, and !<cmd> are notably missing.  This commit implements
these in the same fashion as !!.  Note that this remains far from a
complete implementation of the bash bang command syntax, but it covers
some of the most common commands and is hopefully a step in the right
direction.