git: add pre-commit dependency
This also adds documentation for how to install this git pre-commit hook
which will run black before committing. The only downside to doing that
is the, perhaps confusing, initial run which will clone the black
repository before a dev's first commit.

Running on the command-line will be quite obvsious since the default
output shows what pre-commit is doing. But running in a gui or
editor (e.g. magit) will hide that. Viewer discretion is advised.