Allow disabling individual items in gitstatus prompt
Setting $XONSH_GITSTATUS_* to an empty string lead to the number of files
still being displayed, without any way to see to which category those
belonged. It might even display "11", which could be "1 changed file" and
"1 staged file" or "11 staged files".

Changing the logic to hide the entire entry if the corresponding
$XONSH_GITSTATUS_* setting is empty allows for an easy way to disable some
of the categories.