remove dependency on 'random` library
I wonder what the use case was for a 'random' shell type feature,
nevertheless removing it will reduce start-up time and memory usage.
Apart from this obscure feature, there are no other uses of `import random`
in the main app.

This patch also includes a change in generation of choices for the `--shell-type`
argument. Prior to this patch the `--help` output looked like this:

```
  --shell-type {b,best,d,dumb,ptk,ptk1,ptk2,prompt-toolkit,prompt_toolkit,prompt-toolkit1,prompt-toolkit2,rand,random,rl,readline}
```

After this patch it looks like this:

```
  --shell-type {best,dumb,prompt_toolkit,prompt_toolkit1,prompt_toolkit2,readline}
```

I've changed it since it would still contain 'random' option otherwise,
but I think it also improves the readability of the `--help` output.

And for completeness sake, this feature was introduced in a9c606b8252c3b8ad50ac1735e32528940c9ed5e.