make comments easier to comprehend
Make comments in the test_multiline_...-testcases easier to comprehend
by using correct grammar and spelling and by stating more relevant aspects.

Comment on test parameters with inline instead of block comments,
as this makes it easier to associate code lines with comments,
without having to resort to additional blank lines for separation.