Perform path completion filtering based on expanded path (#4115)
* Perform path completion filtering based on expanded path

This fixes issues with the cd and rmdir completers not working
correctly due to os.path.isdir returning false for paths like
'~/example'.

* Avoid running subsequent completers for cd/rmdir

Raise StopIteration in the cd and rmdir completers if nothing is
found. This prevents the later completers like the generic 'path'
completer from running and returning incorrect results.