Do not repeat commands when searching them in history (using up arrow)
If you execute the same command N times, it appears also N times when using up arrow to search previous commands. It is more convenient for commands not to be repeated during this type of search (as, for example, in fish shell).
I have used a python set to solve this issue, although there may be a better solution.