DOC: Autoreformat most of the docs.
this uses https://github.com/Carreau/velin to try to autoreformat all
docstrings in xonsh/*.py

In particular in Parameters section the space before the colon is
necessary for numpydoc to properly find the name and the type.

When no type is there numpydoc says that colon should be removed.

This also try to find if there are typo in parameter names and fix them
when possible.

As a sideeffect of parsing with numpydoc and reconstructing the
docstring from the parsed data some whitespace information is lost; but
that has the advantage of making the whitespace rules uniform
everywhere.