Refactor: replace singleton XSH with stack
Currently it is not possible to have multiple xonsh environments running together. Whilst this is not something many people would be likely to do, it is a symptom of the fact that we have so many global references to XSH. Additionally, there is little encapsulation as the session is a god-object, making it very hard to refactor things. By making the session immutable, and slowly scoping access, we can modularise the code.

This will also make the test suite cleaner, as it will be easier to ensure that tests run in different sessions where required.