Update README.rst (#5513)
Add x-cmd to the projects-that-use-xonsh-or-compatible list

x-cmd is a toolset implemented using posix shell and awk. It is very
small in size and offers many interesting features. Here is a milestone
demo: https://www.x-cmd.com/mod/onsh

<!---

Thanks for opening a PR on xonsh!

Please do this:

1. Include a news file with your PR
(https://xon.sh/devguide.html#changelog).
2. Add the documentation for your feature into `/docs`.
3. Add the example of usage or before-after behavior.
4. Mention the issue that this PR is addressing e.g. `#1234`.

-->

## For community
⬇️ **Please click the 👍 reaction instead of leaving a `+1` or 👍
comment**

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>