switch to utilize environ.py utilities
Switches to utilize environ.py utilities directly.

This could probably be expanded to also make it so that you can actually
specify the name of the environment variable within the declaration of
the variable in environ.py and then adding a way to actually grab the
environment variable from 'env' that does this conversion + default
grabbing for you.

Signed-off-by: Eli Uriegas <seemethere101@gmail.com>